chore(cleanup): refactor import paths to use gui.app module structure
