+++ src/gui/plugins/menu_extension/plugin_homepage_menu_account.py
@@ -3,7 +3,7 @@ from gui.accountpage.accountpage import AccountPage
 from gui.accountpage.reorder_account_widgets_page import (
     ReorderAccountWidgetsWindow,
 )
-from gui.basewindow import BaseWindow
+from gui.app.basewindow import BaseWindow
 
 menu_id = 10
