+++ src/gui/plugins/menu_extension/plugin_homepage_menu_data.py
@@ -1,6 +1,6 @@
 import tkinter as tk
 import utils.data.createdatabase_utils as db_utils
-from gui.basewindow import BaseWindow
+from gui.app.basewindow import BaseWindow
 
 menu_id = 30
